idiokit.http.server: Fix RuntimeErrors caused by inconsistent handling of HTTP version strings.
The problem was found at Crash Test Party 2015-09-03 with Defensics HTTP Server Suite. It triggers with requests with a HTTP version strings which are valid, but contain e.g. leading zeroes in the major and/or minor parts. The request survives through an initial parsing and checks, but later the code performs a strict comparison to strings "HTTP/1.1" and "HTTP/1.0". As it was assumed that only HTTP/1.0 and HTTP/1.1 requests should ever advance to the comparison point a RuntimeError got raised.

Solution was to handle HTTP version data more consistently (with the new httpversion module) and consolidate the decision making based on the version to one place.